tool: abstract out binary image generation and kernel configuration generation
