+++ tool/microkit/src/build.rs
@@ -1,0 +1,554 @@
+use std::{
+    collections::HashMap,
+    fs::{self, metadata},
+    path::{Path, PathBuf},
+};
+
+use crate::{
+    capdl::{
+        allocation::{simulate_capdl_object_alloc_algorithm, CapDLAllocEmulationErrorLevel},
+        build_capdl_spec,
+        initialiser::CapDLInitialiser,
+        packaging::pack_spec_into_initial_task,
+    },
+    elf::ElfFile,
+    loader::Loader,
+    report::write_report,
+    sdf::{SysMemoryRegion, SysMemoryRegionPaddr, SystemDescription},
+    sel4::{emulate_kernel_boot, emulate_kernel_boot_partial, Arch, Config},
+    symbols::patch_symbols,
+    util::{human_size_strict, round_down, round_up},
+    DisjointMemoryRegion, MemoryRegion,
+};
+
+const MAX_BUILD_ITERATION: usize = 3;
+
+// When building for x86, the kernel is copied from the SDK release package to the same
+// directory as the output boot module image, as Multiboot want them as
+// separate images.
+const KERNEL_COPY_FILENAME: &str = "sel4.elf";
+// The `-kernel` argument of 'qemu-system-x86_64' doesn't accept a 64-bit image, so we
+// also copy the 32-bit version that was prepared by build_sdk.py for convenience.
+const KERNEL32_COPY_FILENAME: &str = "sel4_32.elf";
+
+enum ImageOutputType {
+    Binary,
+    Elf,
+    Uimage,
+}
+
+impl ImageOutputType {
+    fn default_from_arch_and_board(arch: &Arch, board_name: &str) -> Self {
+        match board_name {
+            "ariane" | "cheshire" | "serengeti" => ImageOutputType::Elf,
+            _ => match arch {
+                Arch::Aarch64 => ImageOutputType::Binary,
+                Arch::Riscv64 => ImageOutputType::Uimage,
+                Arch::X86_64 => ImageOutputType::Elf,
+            },
+        }
+    }
+
+    fn parse(str: &str, arch: Arch) -> Result<Self, String> {
+        match str {
+            "binary" => match arch {
+                Arch::Aarch64 | Arch::Riscv64 => Ok(ImageOutputType::Binary),
+                Arch::X86_64 => Err(format!(
+                    "building the output image as binary is unsupported for target architecture '{arch}'"
+                )),
+            },
+            "elf" => Ok(ImageOutputType::Elf),
+            "uimage" => match arch {
+                Arch::Riscv64 => Ok(ImageOutputType::Uimage),
+                Arch::X86_64 | Arch::Aarch64 => Err(format!(
+                    "building the output image as uImage is unsupported for target architecture '{arch}'"
+                )),
+            },
+            _ => Err(format!("unknown value '{str}'")),
+        }
+    }
+}
+
+fn get_full_path(path: &Path, search_paths: &Vec<PathBuf>) -> Option<PathBuf> {
+    for search_path in search_paths {
+        let full_path = search_path.join(path);
+        if full_path.exists() {
+            return Some(full_path.to_path_buf());
+        }
+    }
+
+    None
+}
+
+pub fn build(
+    kernel_config: Config,
+    board: &String,
+    config: &String,
+    system: &mut SystemDescription,
+    sdk_dir: &Path,
+    search_paths: Vec<PathBuf>,
+    output: &String,
+    report: &String,
+    capdl_json: Option<&str>,
+    output_type: Option<&str>,
+) -> Result<(), String> {
+    let elf_path = sdk_dir.join("board").join(board).join(config).join("elf");
+    let loader_elf_path = elf_path.join("loader.elf");
+    let kernel_elf_path = elf_path.join("sel4.elf");
+    let monitor_elf_path = elf_path.join("monitor.elf");
+    let capdl_init_elf_path = elf_path.join("initialiser.elf");
+
+    if !elf_path.exists() {
+        return Err(format!(
+            "Error: board ELF directory '{}' does not exist",
+            elf_path.display()
+        ));
+    }
+    if !kernel_elf_path.exists() {
+        return Err(format!(
+            "Error: kernel ELF '{}' does not exist",
+            kernel_elf_path.display()
+        ));
+    }
+    if !monitor_elf_path.exists() {
+        return Err(format!(
+            "Error: monitor ELF '{}' does not exist",
+            monitor_elf_path.display()
+        ));
+    }
+    if !capdl_init_elf_path.exists() {
+        return Err(format!(
+            "Error: CapDL initialiser ELF '{}' does not exist",
+            capdl_init_elf_path.display()
+        ));
+    }
+    let capdl_initialiser_elf = ElfFile::from_path(&capdl_init_elf_path).unwrap_or_else(|e| {
+        eprintln!(
+            "ERROR: failed to parse initialiser ELF ({}): {}",
+            capdl_init_elf_path.display(),
+            e
+        );
+        std::process::exit(1);
+    });
+
+    // Only relevant for ARM and RISC-V.
+    // Determine how much physical memory is available to the kernel after it boots but before dropping
+    // to userspace by partially emulating the kernel boot process. This is useful for two purposes:
+    // 1. To implement setvar region_paddr for memory regions that doesn't specify a phys address, where
+    //    we must automatically select a suitable address inside the Microkit tool.
+    // 2. Post-spec generation sanity checks at a later point to ensure that there are sufficient memory
+    //    to allocate all kernel objects.
+    let (kernel_elf_maybe, available_memory_maybe, kernel_boot_region_maybe) =
+        match kernel_config.arch {
+            Arch::X86_64 => (None, None, None),
+            Arch::Aarch64 | Arch::Riscv64 => {
+                let kernel_elf = ElfFile::from_path(&kernel_elf_path).unwrap_or_else(|e| {
+                    eprintln!(
+                        "ERROR: failed to parse kernel ELF ({}): {}",
+                        kernel_elf_path.display(),
+                        e
+                    );
+                    std::process::exit(1);
+                });
+
+                // Now determine how much memory we have after the kernel boots.
+                let (available_memory, kernel_boot_region) =
+                    emulate_kernel_boot_partial(&kernel_config, &kernel_elf);
+                (
+                    Some(kernel_elf),
+                    Some(available_memory),
+                    Some(kernel_boot_region),
+                )
+            }
+        };
+
+    let monitor_elf = ElfFile::from_path(&monitor_elf_path).unwrap_or_else(|e| {
+        eprintln!(
+            "ERROR: failed to parse monitor ELF ({}): {}",
+            monitor_elf_path.display(),
+            e
+        );
+        std::process::exit(1);
+    });
+
+    // This list refers to all PD ELFs as well as the Monitor ELF.
+    // The monitor is very similar to a PD so it is useful to pass around
+    // a list like this.
+    let mut system_elfs = Vec::with_capacity(system.protection_domains.len());
+    // Get the elf files for each pd:
+    for pd in &system.protection_domains {
+        match get_full_path(&pd.program_image, &search_paths) {
+            Some(path) => {
+                let path_for_symbols = pd
+                    .program_image_for_symbols
+                    .as_ref()
+                    .map(|path_suffix| {
+                        get_full_path(path_suffix, &search_paths).ok_or_else(|| {
+                            format!(
+                                "unable to find program image for symbols: '{}'",
+                                path_suffix.display()
+                            )
+                        })
+                    })
+                    .transpose()?;
+                match ElfFile::from_split_paths(&path, path_for_symbols.as_deref()) {
+                    Ok(elf) => system_elfs.push(elf),
+                    Err(e) => {
+                        eprintln!(
+                            "ERROR: failed to parse ELF '{}' for PD '{}': {}",
+                            path.display(),
+                            pd.name,
+                            e
+                        );
+                        std::process::exit(1);
+                    }
+                };
+            }
+            None => {
+                return Err(format!(
+                    "unable to find program image: '{}'",
+                    pd.program_image.display()
+                ))
+            }
+        }
+    }
+
+    // The monitor is just a special PD
+    system_elfs.push(monitor_elf);
+
+    let mut capdl_initialiser = CapDLInitialiser::new(capdl_initialiser_elf);
+
+    // Now build the capDL spec and final image. We may need to do this in >1 iterations on ARM and RISC-V
+    // if there are Memory Regions without a paddr but subject to setvar region_paddr.
+    let mut iteration = 0;
+    let mut spec_need_refinement = true;
+    let mut system_built = false;
+    while spec_need_refinement && iteration < MAX_BUILD_ITERATION {
+        spec_need_refinement = false;
+
+        // Patch all the required symbols in the Monitor and PDs according to the Microkit's requirements
+        if let Err(err) = patch_symbols(&kernel_config, &mut system_elfs, &system) {
+            eprintln!("ERROR: {err}");
+            std::process::exit(1);
+        }
+
+        let mut spec_container = build_capdl_spec(&kernel_config, &mut system_elfs, &system)?;
+        pack_spec_into_initial_task(
+            &kernel_config,
+            config,
+            &spec_container,
+            &system_elfs,
+            &mut capdl_initialiser,
+        );
+
+        match kernel_config.arch {
+            Arch::X86_64 => {
+                // setvar region_paddr not supported on this architecture nor can we emulate the
+                // kernel boot process to statically check for issues due to unknown memory map, so nothing to do.
+                // Write out the capDL initialiser as an ELF boot module and we are done.
+            }
+            Arch::Aarch64 | Arch::Riscv64 => {
+                // Now that we have the CapDL initialiser ELF with embedded spec,
+                // we can determine exactly how much memory will be available statically when the kernel
+                // drops to userspace on ARM and RISC-V. This allow us to sanity check that:
+                // 1. There are enough memory to allocate all the objects required in the spec.
+                // 2. All frames with a physical attached reside in legal memory (device or normal).
+                // 3. Objects can be allocated from the free untyped list. For example, we detect
+                //    situations where you might have a few frames with size bit 12 to allocate but
+                //    only have untyped with size bit <12 remaining.
+                // This also allow the tool to automatically pick physical address of Memory Regions with out
+                // an explicit paddr in SDF but are subject to setvar region_paddr.
+
+                // Determine how much memory the CapDL initialiser needs.
+                let initialiser_vaddr_range = capdl_initialiser.image_bound();
+                let initial_task_size = initialiser_vaddr_range.end - initialiser_vaddr_range.start;
+
+                // Reuse data from the partial kernel boot emulation previously done.
+                // .clone() as we need to mutate this for every iteration.
+                let mut available_memory = available_memory_maybe.clone().unwrap();
+                let kernel_boot_region = kernel_boot_region_maybe.unwrap();
+
+                // The kernel relies on the initial task region being allocated above the kernel
+                // boot/ELF region, so we have the end of the kernel boot region as the lower
+                // bound for allocating the reserved region.
+                let initial_task_phys_base =
+                    available_memory.allocate_from(initial_task_size, kernel_boot_region.end);
+
+                let Some(initial_task_phys_base) = initial_task_phys_base else {
+                    // Unlikely to happen on Microkit-supported platforms with multi gigabytes memory.
+                    // But printing a helpful error in case we do run into this problem.
+                    eprintln!(
+                        "ERROR: cannot allocate memory for the initialiser, contiguous physical memory region of size {} not found", human_size_strict(initial_task_size)
+                    );
+                    eprintln!("ERROR: physical memory regions the initialiser can be placed at:");
+                    for region in available_memory.regions {
+                        eprintln!(
+                            "       [0x{:0>12x}..0x{:0>12x}), size: {}",
+                            region.base,
+                            region.end,
+                            human_size_strict(region.size())
+                        );
+                    }
+                    std::process::exit(1);
+                };
+
+                capdl_initialiser.set_phys_base(initial_task_phys_base);
+                let initial_task_phys_region = MemoryRegion::new(
+                    initial_task_phys_base,
+                    initial_task_phys_base + initial_task_size,
+                );
+                let user_image_virt_region = MemoryRegion::new(
+                    capdl_initialiser.elf.lowest_vaddr(),
+                    initialiser_vaddr_range.end,
+                );
+
+                // With the initial task region determined the kernel boot can be emulated in full. This provides
+                // the boot info information (containing untyped objects) which is needed for the next steps
+                let kernel_boot_info = emulate_kernel_boot(
+                    &kernel_config,
+                    kernel_elf_maybe.as_ref().unwrap(),
+                    initial_task_phys_region,
+                    user_image_virt_region,
+                );
+
+                if iteration == 0 {
+                    // On the first iteration where the spec have not been refined, simulate the capDL allocation algorithm
+                    // to double check that all kernel objects of the system as described by SDF can be successfully allocated.
+                    if !simulate_capdl_object_alloc_algorithm(
+                        &mut spec_container,
+                        &kernel_boot_info,
+                        &kernel_config,
+                        CapDLAllocEmulationErrorLevel::PrintStderr,
+                    ) {
+                        eprintln!("ERROR: could not allocate all required kernel objects. Please see report for more details.");
+                        std::process::exit(1);
+                    }
+                } else {
+                    // Do the same thing for further iterations, at this point the simulation won't fail *except* for when we have picked a
+                    // bad address for Memory Regions subject to setvar region_paddr. This can happen because after we have
+                    // picked the address, we will update spec and patch it into the program's frame. Which will causes the
+                    // spec to increase in size as the frames' data are compressed. So if the simulation fail, we need to
+                    // pick another address as we now have a better idea of how large the spec is.
+
+                    // This is highly unlikely to happen unless the spec size increase causes the initial task size to cross
+                    // a 4K page boundary.
+                    if !simulate_capdl_object_alloc_algorithm(
+                        &mut spec_container,
+                        &kernel_boot_info,
+                        &kernel_config,
+                        CapDLAllocEmulationErrorLevel::Suppressed,
+                    ) {
+                        // Encountered a problem, pick a better address.
+                        for tool_allocate_mr in system.memory_regions.iter_mut().filter(|mr| {
+                            matches!(mr.phys_addr, SysMemoryRegionPaddr::ToolAllocated(_))
+                        }) {
+                            tool_allocate_mr.phys_addr = SysMemoryRegionPaddr::ToolAllocated(None);
+                        }
+                        spec_container.expected_allocations = HashMap::new();
+                    }
+                }
+
+                // Now pick a physical address for any memory regions that are subject to setvar region_paddr.
+                // Doing something a bit unconventional here: converting the list of untypeds back to a DisjointMemoryRegion
+                // to give us a view of physical memory available after the kernel drops to user space.
+                // I.e. available memory after the initial task have been created.
+                {
+                    let mut available_user_memory = DisjointMemoryRegion::default();
+                    for ut in kernel_boot_info
+                        .untyped_objects
+                        .iter()
+                        .filter(|ut| !ut.is_device)
+                    {
+                        // Only take untypeds that can at least fit a page because some have been used to back the initial task's
+                        // kernel object such as TCB, endpoint etc.
+                        let start = round_up(ut.base(), kernel_config.minimum_page_size);
+                        let end = round_down(ut.end(), kernel_config.minimum_page_size);
+                        if end > start {
+                            // will be automatically merged
+                            available_user_memory.insert_region(ut.base(), ut.end());
+                        }
+                    }
+
+                    // Then take away any memory ranges occupied by Memory Regions with a paddr specified in SDF.
+                    for mr in system.memory_regions.iter() {
+                        if let SysMemoryRegionPaddr::Specified(sdf_paddr) = mr.phys_addr {
+                            let mr_end = sdf_paddr + mr.size;
+
+                            // MR may be device memory, which isn't covered in available_user_memory.
+                            let is_normal_mem =
+                                available_user_memory.regions.iter().any(|region| {
+                                    sdf_paddr >= region.base
+                                        && sdf_paddr < region.end
+                                        && mr_end <= region.end
+                                });
+                            if is_normal_mem {
+                                available_user_memory.remove_region(sdf_paddr, sdf_paddr + mr.size);
+                            }
+                        }
+                    }
+
+                    let mut tool_allocated_mrs = Vec::new();
+                    for (mr_id, tool_allocate_mr) in system
+                        .memory_regions
+                        .iter_mut()
+                        .enumerate()
+                        .filter(|(_, mr)| {
+                            matches!(mr.phys_addr, SysMemoryRegionPaddr::ToolAllocated(None))
+                        })
+                    {
+                        spec_need_refinement = true;
+
+                        let target_paddr = available_user_memory
+                            .allocate(tool_allocate_mr.size, tool_allocate_mr.page_size);
+                        if target_paddr.is_none() {
+                            eprintln!("ERROR: cannot auto-select a physical address for MR {} because there are no contiguous memory region of sufficient size.", tool_allocate_mr.name);
+                            eprintln!("ERROR: MR {} needs to be physically contiguous as it is a subject of a setvar region_paddr.", tool_allocate_mr.name);
+                            if !tool_allocated_mrs.is_empty() {
+                                eprintln!("Previously auto-allocated memory regions:");
+                                for allocated_mr_id in tool_allocated_mrs {
+                                    let allocated_mr: &SysMemoryRegion =
+                                        &system.memory_regions[allocated_mr_id];
+                                    eprintln!(
+                                        "name = '{}', paddr = 0x{:0>12x}, size = 0x{:0>12x}",
+                                        allocated_mr.name,
+                                        allocated_mr.paddr().unwrap(),
+                                        allocated_mr.size
+                                    );
+                                }
+                            }
+                            eprintln!("available physical memory regions:");
+                            for region in available_user_memory.regions {
+                                eprintln!(
+                                    "[0x{:0>12x}..0x{:0>12x}), size: {}",
+                                    region.base,
+                                    region.end,
+                                    human_size_strict(region.size())
+                                );
+                            }
+                            std::process::exit(1);
+                        }
+                        tool_allocated_mrs.push(mr_id);
+                        tool_allocate_mr.phys_addr =
+                            SysMemoryRegionPaddr::ToolAllocated(target_paddr);
+                    }
+                }
+
+                // Patch the list of untypeds we used to simulate object allocation into the initialiser.
+                // At runtime the intialiser will validate what we simulated against what the kernel gives it. If they deviate
+                // we will have problems! For example, if we simulated with more memory than what's actually available, the initialiser
+                // can crash.
+                capdl_initialiser.add_expected_untypeds(&kernel_boot_info.untyped_objects);
+            }
+        };
+
+        if !spec_need_refinement {
+            // All is well in the universe, write the image out.
+            println!(
+                "MICROKIT|CAPDL SPEC: number of root objects = {}, spec footprint = {}",
+                spec_container.spec.objects.len(),
+                human_size_strict(
+                    capdl_initialiser
+                        .spec_metadata()
+                        .as_ref()
+                        .unwrap()
+                        .spec_size
+                ),
+            );
+            let initialiser_vaddr_range = capdl_initialiser.image_bound();
+            println!(
+                "MICROKIT|INITIAL TASK: memory size = {}",
+                human_size_strict(initialiser_vaddr_range.end - initialiser_vaddr_range.start),
+            );
+
+            let image_out_path = Path::new(output);
+
+            match kernel_config.arch {
+                Arch::X86_64 => match capdl_initialiser.elf.reserialise(image_out_path) {
+                    Ok(size) => {
+                        // Copy the kernel to the build directory as well so users doesn't have to dig through the SDK.
+                        if let Err(copy_err) = fs::copy(
+                            &kernel_elf_path,
+                            image_out_path.parent().unwrap().join(KERNEL_COPY_FILENAME),
+                        ) {
+                            eprintln!("ERROR: couldn't copy the kernel to image's output directory: {copy_err}");
+                            std::process::exit(1);
+                        }
+                        if let Err(copy_err) = fs::copy(
+                            kernel_elf_path
+                                .parent()
+                                .unwrap()
+                                .join(KERNEL32_COPY_FILENAME),
+                            image_out_path
+                                .parent()
+                                .unwrap()
+                                .join(KERNEL32_COPY_FILENAME),
+                        ) {
+                            eprintln!("ERROR: couldn't copy the 32-bit kernel to image's output directory: {copy_err}");
+                            std::process::exit(1);
+                        }
+                        println!(
+                            "MICROKIT|BOOT MODULE: image file size = {}",
+                            human_size_strict(size)
+                        );
+                    }
+                    Err(err) => {
+                        eprintln!("ERROR: couldn't write the boot module to filesystem: {err}");
+                        std::process::exit(1);
+                    }
+                },
+                Arch::Aarch64 | Arch::Riscv64 => {
+                    let loader = Loader::new(
+                        &kernel_config,
+                        Path::new(&loader_elf_path),
+                        kernel_elf_maybe.as_ref().unwrap(),
+                        &capdl_initialiser.elf,
+                        capdl_initialiser.phys_base.unwrap(),
+                        &initialiser_vaddr_range,
+                    );
+
+                    let image_output_type = match output_type {
+                        Some(s) => ImageOutputType::parse(&s, kernel_config.arch).unwrap_or(
+                            ImageOutputType::default_from_arch_and_board(
+                                &kernel_config.arch,
+                                board,
+                            ),
+                        ),
+                        None => {
+                            ImageOutputType::default_from_arch_and_board(&kernel_config.arch, board)
+                        }
+                    };
+
+                    match image_output_type {
+                        ImageOutputType::Binary => loader.write_image(image_out_path),
+                        ImageOutputType::Elf => loader.write_elf(image_out_path),
+                        ImageOutputType::Uimage => loader.write_uimage(image_out_path),
+                    };
+
+                    println!(
+                        "MICROKIT|LOADER: image file size = {}",
+                        human_size_strict(metadata(image_out_path).unwrap().len())
+                    );
+                }
+            };
+
+            if let Some(capdl_json) = capdl_json {
+                let serialised = serde_json::to_string_pretty(&spec_container.spec).unwrap();
+                fs::write(capdl_json, &serialised).unwrap();
+            };
+
+            write_report(&spec_container, &kernel_config, report);
+            system_built = true;
+            break;
+        } else {
+            // Some memory regions have had their physical address updated, rebuild the spec.
+            iteration += 1;
+        }
+    }
+
+    if !system_built {
+        // Cannot build a reasonable spec, absurd.
+        // Only reachable when there are setvar region_paddr that we keep selecting the wrong address.
+        panic!("ERROR: fatal, failed to build system in {iteration} iterations");
+    }
+    Ok(())
+}
